fix: webauthn sign fix

+++ unipay/confirm_status.go
@@ -133,7 +133,7 @@ type ReqCouponOrderCreate struct {
 	core.ChainTypeAddress
 	Account   string         `json:"account" binding:"required"`
 	TokenId   tables.TokenId `json:"token_id" binding:"required"`
-	Num       int            `json:"num" binding:"min=1,max=10000"`
+	Num       int64          `json:"num" binding:"min=1,max=10000"`
 	Cid       string         `json:"cid" binding:"required"`
 	Name      string         `json:"name" binding:"required"`
 	Note      string         `json:"note"`
